second stab at find device dialog
