+++ imls-raspberry-pi/cmd/wifi-hardware-search-windows/find-device.go
@@ -1,0 +1,27 @@
+// go:build windows
+package main
+
+import (
+	"fmt"
+
+	"golang.org/x/sys/windows"
+	"gsa.gov/18f/internal/wifi-hardware-search/models"
+	"gsa.gov/18f/internal/wifi-hardware-search/search"
+)
+
+func main() {
+	device := new(models.Device)
+	device.Exists = false
+	search.FindMatchingDevice(device)
+	title := windows.StringToUTF16Ptr("IMLS: compatible wifi device query")
+	if device.Exists {
+		message := windows.StringToUTF16Ptr(fmt.Sprintf("found a compatible wifi device: %s (%s) [%s]",
+			device.Logicalname,
+			device.Description,
+			device.Vendor))
+		windows.MessageBox(0, message, title, windows.MB_OK)
+	} else {
+		message := windows.StringToUTF16Ptr("no compatible wifi device was found")
+		windows.MessageBox(0, message, title, windows.MB_ICONWARNING)
+	}
+}
